+++ move.h
@@ -1,13 +1,14 @@
 #ifndef MOVE_H
 #define MOVE_H
+#include <QtGlobal>
 
 class Move
 {
-    int from;
-    int to;
+    qint16 from;
+    qint16 to;
 
 public:
-    Move(int, int);
+    Move(qint16, qint16);
     Move();
 
     int getFrom() const {return from;}
